+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Added a sample project using `ExParamsSchema.Handler` and JSON Schema validation.
+
 ## [0.1.0] - 2026-08-11
 
 ### Added